Update dashboard API
