+++ db/contract.go
@@ -27,6 +27,7 @@ type IContract interface {
 	UpdateKRCTotalSupply(ctx context.Context, krcTokenAddress, totalSupply string) error
 	Contracts(ctx context.Context, filter *types.ContractsFilter) ([]*types.Contract, uint64, error)
 
+	CountContracts(ctx context.Context) (int64, error)
 	AllContracts(ctx context.Context) ([]*types.Contract, error)
 	ContractByType(ctx context.Context, contractType string) ([]*types.Contract, error)
 	UpsertSMCABIByType(ctx context.Context, smcType, abi string) error
@@ -35,6 +36,15 @@ type IContract interface {
 	// Remove
 	RemoveContract(ctx context.Context, contractAddress string) error
 	RemoveContracts(ctx context.Context) error
+}
+
+func (m *mongoDB) CountContracts(ctx context.Context) (int64, error) {
+	total, err := m.wrapper.C(cContract).Count(bson.M{})
+	if err != nil {
+		return 0, err
+	}
+
+	return total, nil
 }
 
 func (m *mongoDB) InsertContract(ctx context.Context, contract *types.Contract, addrInfo *types.Address) error {
